# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateTransactionRequest:
    properties:
      customer_id:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.TransactionItemRequest'
        type: array
      payment_method:
        description: 1=Cash, 2=Midtrans
        enum:
        - 1
        - 2
        type: integer
      tenant_id:
        type: string
    required:
    - items
    - payment_method
    - tenant_id
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 5
        type: string
      role_ids:
        items:
          type: string
        type: array
      tenant_id:
        type: string
    required:
    - email
    - name
    - password
    - role_ids
    type: object
  dto.PayTransactionRequest:
    properties:
      amount_tendered:
        type: number
    required:
    - amount_tendered
    type: object
  dto.PayTransactionResponse:
    properties:
      amount_tendered:
        type: number
      change:
        type: number
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.TransactionItemResponse'
        type: array
      payment_method:
        type: integer
      payment_status:
        type: integer
      total_amount:
        type: number
    type: object
...
  dto.TransactionItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.TransactionItemResponse:
    properties:
      price:
        type: number
      product_id:
        type: string
      quantity:
        type: integer
      subtotal:
        type: number
    type: object
  dto.TransactionResponse:
    properties:
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/dto.TransactionItemResponse'
        type: array
      payment_method:
        description: 1=Cash, 2=Midtrans
        type: integer
      payment_status:
        type: integer
      tenant_id:
        type: string
      total_amount:
        type: number
      user_id:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role_ids:
        items:
          type: string
        type: array
      tenant_id:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Point of Sale (POS) application with multi-tenant support.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: POS Backend API
  version: "1.0"
paths:
...
  /me:
    get:
      description: Retrieve profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - User
...
  /transactions/:
    post:
      consumes:
      - application/json
      description: Create a new transaction. `tenant_id` required; `user_id` is taken
        from the authenticated user.
      parameters:
      - description: Create Transaction Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TransactionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create transaction
      tags:
      - Transaction
  /transactions/{id}/pay:
    post:
      consumes:
      - application/json
      description: Process payment for a transaction by ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PayTransactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PayTransactionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Pay transaction
      tags:
      - Transaction
  /transactions/{id}/receipt:
    get:
      description: Generate and download PDF receipt for a transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: PDF receipt
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get transaction receipt PDF
      tags:
      - Transaction
  /users/:
    post:
      consumes:
      - application/json
      description: Create a new user with role assignment
      parameters:
      - description: Create User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new user
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"